requirements.txt
fastapi==0.115.12           
uvicorn==0.34.2             
SQLAlchemy==2.0.40          
psycopg2-binary==2.9.10     
python-dotenv==1.1.0        
passlib[bcrypt]==1.7.4      
python-jose==3.4.0          
pydantic==2.11.4            
alembic==1.15.2             
httpx==0.24.1              
pytest==7.4.0             
pytest-asyncio==0.21.0    
pytest-cov==4.1.0        
pytest-mock==3.12.0      
httpx-mock==0.20.0
